Add secret halloween mare map prototype.
Cannot be compiled as the layout is not yet fully built.
If I can't finish this a few days before halloween, then it'll stay unfinished until next year.
Or later turned into a non-halloween map, who knows.

This commit also removes all the remaining map vmfs from the main branch that should not be in the dev branch.